fix: no exit msg in pipe:
