+++ execute/internal/execute_utils.c
@@ -2,7 +2,8 @@
 #include "../builtin/builtin.h"
 #include "execute_internal.h"
 
-void	new_executor(t_executor **e, t_env_var **env_vars, t_ast_node *root, bool is_interactive)
+void	new_executor(t_executor **e, t_env_var **env_vars,
+					 t_ast_node *root, bool is_interactive)
 {
 	*e = x_malloc(sizeof(**e));
 	(*e)->exit_status = EXIT_SUCCESS;
@@ -22,4 +23,43 @@ bool	is_execute_condition(int condition, int exit_status)
 	if (condition == CONDITION_NL)
 		return (true);
 	return (false);
+}
+
+bool	is_dir(const char *path)
+{
+	struct stat	stat_;
+
+	x_stat(path, &stat_);
+	if ((stat_.st_mode & S_IFMT) == S_IFDIR)
+		return (true);
+	else
+		return (false);
+}
+
+void	delete_execute_list(void *element, t_execute_list_type type)
+{
+	if (!element)
+		return ;
+	else if (type == T_SIMPLE_COMMAND)
+	{
+		free(((t_simple_command *)element)->argv);
+		close(((t_simple_command *)element)->r_out);
+		close(((t_simple_command *)element)->r_in);
+	}
+	else if (type == T_COMPOUND_LIST)
+	{
+		delete_execute_list(((t_compound_list *)element)->pipeline, T_PIPELINE);
+		delete_execute_list(
+			((t_compound_list *)element)->next, T_COMPOUND_LIST);
+	}
+	else if (type == T_SUBSHELL)
+		delete_execute_list(
+			((t_subshell *)element)->compound_list, T_COMPOUND_LIST);
+	else if (type == T_PIPELINE)
+	{
+		delete_execute_list(
+			((t_pipeline *)element)->command, ((t_pipeline *)element)->type);
+		delete_execute_list(((t_pipeline *)element)->next, T_PIPELINE);
+	}
+	free(element);
 }
